+++ packages/file-types/csv/src/csv-loaders/ObsSpotsCsv.js
@@ -24,7 +24,7 @@ export default class ObsSpotsCsvLoader extends CsvLoader {
 
     const coordinationValues = {
       spotLayer: CL({
-        obsType: 'spot',
+        obsType: this.coordinationValues?.obsType ?? 'spot',
         // obsColorEncoding: 'spatialLayerColor',
         // spatialLayerColor: [255, 255, 255],
         spatialLayerVisible: true,