+++ packages/file-types/zarr/src/anndata-loaders/ObsPointsAnndataLoader.js
@@ -26,7 +26,7 @@ export default class ObsPointsAnndataLoader extends AbstractTwoStepLoader {
     const { path } = this.options;
     const coordinationValues = {
       pointLayer: CL({
-        obsType: 'point',
+        obsType: this.coordinationValues?.obsType ?? 'point',
         // obsColorEncoding: 'spatialLayerColor',
         // spatialLayerColor: [255, 255, 255],
         spatialLayerVisible: true,