+++ packages/styles/src/index.ts
@@ -93,6 +93,7 @@ export { default as ArrowLeft } from '@mui/icons-material/ArrowLeft';
 export { default as CenterFocusStrong } from '@mui/icons-material/CenterFocusStrong';
 export { default as Close } from '@mui/icons-material/Close';
 export { default as CloudDownload } from '@mui/icons-material/CloudDownload';
+export { default as MenuBook } from '@mui/icons-material/MenuBook';
 export { default as ExpandLess } from '@mui/icons-material/ExpandLess';
 export { default as ExpandMore } from '@mui/icons-material/ExpandMore';
 export { default as Help } from '@mui/icons-material/Help';
@@ -119,6 +120,7 @@ export { default as ViewCompact } from '@mui/icons-material/ViewCompact';
 export { default as LeakAdd } from '@mui/icons-material/LeakAdd';
 export { default as Warning } from '@mui/icons-material/Warning';
 export { default as Palette } from '@mui/icons-material/Palette';
+export { default as ClearIcon } from '@mui/icons-material/Clear';
 
 // Data Grid
 export { DataGrid } from '@mui/x-data-grid/DataGrid';