+++ packages/utils/zarr-utils/src/index.ts
@@ -7,3 +7,7 @@ export {
 } from './normalize.js';
 export { createStoreFromMapContents } from './base64-store.js';
 export { createGetRange } from './base-getrange.js';
+export {
+  flattenOmeAttrs,
+  getOmeNgffVersion,
+} from './ome-ngff-version.js';