+++ sites/docs/src/pages/_ScreenshotImage.js
@@ -6,6 +6,7 @@ export default function ScreenshotImage(props) {
   const {
     alt = '',
     filename,
+    noTitle = false,
   } = props;
   const { colorMode } = useColorMode();
   const isDarkTheme = (colorMode === 'dark');
@@ -13,8 +14,8 @@ export default function ScreenshotImage(props) {
   const baseUrl = useBaseUrl('/img/screenshots');
   return (
     <div>
-      <img src={`${baseUrl}/${theme}/${filename}`} alt={alt} title={alt} />
-      <h3 style={{ textAlign: 'center' }}>{alt}</h3>
+      <img src={`${baseUrl}/${theme}/${filename}`} alt={alt} title={alt} style={{ width: '100%', display: 'block' }} />
+      {!noTitle && <h3 style={{ textAlign: 'center' }}>{alt}</h3>}
     </div>
   );
 }